refactor to helper
